Catenary adjustment to support very sloped seabed
This edit adds support for cases where a line end A is on the
seabed and end B is BELOW end A but off the seabed because the
seabed has a large downward slope. I.e. tan(alpha) < zf/xf where
both are negative. Because seabed contact is only set up for end
A, this edit allows - ZF values in these cases rather than
doing the reverse trick that swaps ends A and B in catenary.