Reformulation to compute dynamic matrices of lines separately
Previously, each call of system.getSystemDynamicMatrices() (and other equivalent functions in the body and point classes) would recompute the dynamic matrices of the line (inertia, added mass, damping, and stiffness). That required the sea state and the motions of the line nodes at each call.

Now, we need to precompute the dynamic matrices of each line using line.updateLumpedMass(). The advantages are:
1. Fewer computations
2. It is simpler to integrate the lumped mass approach in RAFT because now it's easier to account for motions of the fairlead

Also, the only matrix that requires the sea state and motions of the line is the damping matrix. So this commit reformulates the system, body, and line usage of the lumped mass approach to be able to provide inertia, added mass, and stiffness matrices without knowledge of the sea state or motions of the line.  I will add some examples, but you just need to call system.updateSystemDynamicMatrices() and then system.getSystemDynamicMatrices()