Using Guyan reduction in the lumped mass approach
In commit 9b175e06edf7c0d046b2090374d92e67ba8bdcdd we adopted Guyan reduction instead of the double-inversion method to compute system stiffness in system.getCoupledStiffnessA(). Doing the same in the methods that compute dynamic matrices (stiffness, mass, damping, and added mass) using the lumped mass approach.

This isn't working fully as expected yet. The stiffness matrix returned by line.getCoupledDynamicMatrices() still does not match line.KA, line.KB, and line.KBA computed with the quasi-static method very well. For some reason, we still need to remove the nodes on the floor to get a good agreement with the quasi-static results. We will investigate that in the future.

Also, in the lumped mass approach, the guyan_reduce function is being applied to all dynamic matrices in the same way, but the damping and inertia matrices would need to have a special treatment. Will get back to this once the stiffness matrix is working as expected.
235299